packages/sample/create_ghibli_image/requirements.txt
openai==1.70.0
pillow==11.1.0
boto3==1.37.26